expose necessary functions to python, and add round-way tests for
function schema str() and parsing functions.
We iterate over all the registered function schemas and get the string,
then parse the string. We compare the schema generated from parsing with
the original one, and make sure they are equal.

Differential Revision: [D16435471](https://our.internmc.facebook.com/intern/diff/D16435471/)

+++ aten/src/ATen/TensorUtils.cpp
@@ -341,7 +341,7 @@ c10::optional<std::vector<int64_t>> computeStride(
   std::vector<int64_t> newstride(newshape.size());
   if (numel == 0) {
     for (int64_t view_d = newshape.size() - 1; view_d >= 0; view_d--) {
-      if (view_d == newshape.size() - 1) {
+      if (view_d == (int64_t)(newshape.size() - 1)) {
         newstride[view_d] = 1;
       } else {
         newstride[view_d] =
