+++ caffe2/predictor/emulator/data_filler.h
@@ -105,7 +105,7 @@ class DataRandomFiller : public Filler {
       int input_index,
       const std::vector<std::vector<int64_t>>& input_dims) {
     Workspace ws;
-    for (size_t i = 0; i < op_def.input_size(); ++i) {
+    for (int i = 0; i < op_def.input_size(); ++i) {
       // CreateOperator requires all input blobs present
       ws.CreateBlob(op_def.input(i));
     }
